Update patch set 5
Patch Set 5: Code-Review-1

The problem you are trying solve with this patch is one of contention.  Unfortunately this solution adds more contention to the situation.  We used to have this problem quite a lot when our mirrors had older versions of bit on them, versions which took a long time to process repos with many refs.  If this solution were applied to such a situation it would be a disaster.  We finally dealt with our situation be reducing contention: we fixed our mirrors and we added a replication delay which is at least as long as the replication takes.  We now very rarely see these errors.  Now they are still possible, and a better fix from the gerrit side would be to have two queues per project (the extra queue would hold pending requests) to prevent any project from ever replicating from two separate threads to the same mirror.  This would be similar to increasing the replication delay, but it would be more reliable and allow the replication delay to be reduced for the non contended case.

I have often proposed the fix you have here, but the above reason and the permanent failure case have stopped me from pushing forward with it because it felt unwise to me.  That being said, I still wish I had it, but I am very concerned that for the average gerrit installation it could be a mistake.  Perhaps it should be a configuration option disabled by default?  Perhaps the docs should suggest not turning it on until the replication delay has been tuned to prevent failures in the normal case?  Perhaps there should be a retry limit to prevent the infinite loop case (of course, it would be better if you could figure out how to detect that case, it must be possible since parsing the output of ls-remote would reveal the file/directory conflict).

Patch-set: 5
Label: Code-Review=-1