Update patch set 5
Patch Set 5:

Hi, Martin. I understand your concerns, even though I personally think the benefits of the patch out-weights the disadvantages.

> Perhaps there should be a retry limit to prevent the infinite loop case (of course, it would be better if you could figure out how to detect that case, it must be possible since parsing the output of ls-remote would reveal the file/directory conflict).

That's exactly how the code works today. So are we OK with this patch? Is this not better than the alternative than having failed replication pushes?

Patch-set: 5
Label: Code-Review=0