Catch RemoteRepositoryException while replicating to remote server
When using git:// protocol to replicate, if the remote repository has
some problems, such as access denied, repository not exported or not
found, jgit will throw RemoteRepositoryException.

If this exception is not explicitly caught, it will instead be caught as
TransportException, and the replication will be rescheduled.  However in
this case, rescheduling is useless.

Change-Id: I6ed65656664bac0b4c55c9112c42c4812b98a2ea