Update patch set 6
Patch Set 6:

This patch properly removes a job from the pending queue when a collision is detected.

I don't see a good way to refactor the PushOne.canceled manipulation, and Destination.requestRunway() will still need to check the canceled status of its argument, to determine whether it needs to be removed from the pending queue.  So this is as good as it gets for now.

Patch-set: 6