Update patch set 5
Patch Set 5:

(24 comments)

Mostly nits.

If not in --wait mode, whether or not the plugin's start command prints

  Replicate $project_name to $host, SUCCEEDED!
  [...]
  ----------------------------------------------
  Replication completed successfully!

depends solely on timing. I do not so much care, whether or not it
prints anything, but I'd prefer deterministic behaviour. The same
holds true for all the wrapped logs.

Patch-set: 5
Label: Verified=0