Update patch set 7
Patch Set 7:

Sorry, I did not notice your comment, when rebasing Patch Set 5 for
the changes in change 37711.

> What's the use case for this? I can't see it being a huge overhead
> that you have to put ${name} in the config file even if you're only
> replicating one project.

The benefit is that this change allows you to replicate

  some/project

to

  completely/different/name

on the remote.

(I now made that more explicit in the commit message)

But why would that be needed?

Some of our projects (say 'random-awesome-tool') were started on
github, when not yet using gerrit. They've got a following and
contributors there (under that name), which we do not want to loose.

However, in our gerrit, 'random-awesome-tool' was imported under a
different name, say 'divisionB/random-awesome-tool', to meet our
naming scheme in gerrit.

So when replicating our gerrit repos back to github, we want
'divisionB/random-awesome-tool' getting replicated back to the
original 'random-awesome-tool' again.

This change would allow to do that.

Patch-set: 7
Reviewer: Christian Aistleitner <1012155@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0