Introduce remoteNameStyle = basenameOnly
In order to replicate repositories below a certain namespace (e.g. foo/bar/),
the projects parameter allows to filter repositories (projects = foo/bar/*).

With the current implementation, the path on the target server contains the full
path to these repositories (foo/bar/myrepo), which can only be modified in terms
of replacing slashes with dash or underscore using the remoteNameStyle option.

This patch adds a new value "basenameOnly", which only takes the basename of the
repository path (e.g. myrepo when applied to foo/bar/myrepo), which, for example
enables Gerrit user to replicate whole namespaces to a special Github
organization.

Change-Id: Id12780948a4841c054eb76e359f02dae0816d576