Update patch set 11
Patch Set 11:

I think that this change is still valuable.  Failed to locks are still possible if someone modifies a repo behind gerrit's back.  It will also be possible in multi-master setups, so this is very valuable for that!

Also, now that the normal path is fixed to prevent failed-to-locks, I think that it might make sense for this option to be on by default (with the "consolidating refs" change merged, this change should no longer excasberate a loaded server by adding more load)?

Patch-set: 11