Create missing repositories on the remote site
If during the replication of a ref it is found that the repository is
missing on the remote site automatically create it on the remopte
site.

E.g. a repository can be missing on the remote site if the remote
site was not available at the moment when a new project was created
and hence the repository for the new project could not be created on
the remote site.

This feature can also help to  simplify the setup of new remote site.

Whether missing repositories should be created or not can be
configured per destination. By default missing repositories are not
created.

Change-Id: I4e587cdfca09445c9b1c528b2f1edae0944aec68
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>