[RFC] Splitting replication queue from config.
Refactoring to split the responsibility of managing
the replication queue from the actual reading and
processing of the configuration (single responsibility
principle).

This is a preparatory work in order to improve the
way that replication config is managed and loaded.
The aim is to allow a more dynamic and automatic
configuration of replication destinations, without
impacting the current replication logic.

Change-Id: Iad823563e9e41e4ba6bafeef3e993861b2246708