Add port to replication status and display path for local replication
Replication status was only displaying the hostname. It is impossible
to differentiate replication status when 2 urls use the same hostname
with different ports.

Replication status now displays host[:port] and if the replication
url does not have host(file:///), displays the path instead.

Here are some examples of replication url and the replication status
that is displayed:

  -url = git://someHost.com/${name}.git
    Replicate project1 to someHost, Succeeded!

  -url = git://someHost:1234/${name}.git
    Replicate project1 to someHost:1234, Succeeded!

  -url = file:///somePath/git/${name}.git
    Replicate project1 to /somePath/git/project1.git, Succeeded!

Change-Id: Ia86154cb9f26dfdcae37dc9993fe2cd0f7c37838