Update patch set 4
Patch Set 4:

Since the replication plugin is replicating any type of refs, not only refs/changes, we should send event for any ref replicated.

I did not want to upload my change as new patchset since it is not my change but I could if you guys are ok with it. I did it and pushed it as a draft so you could see my change(which is based on this one). I also simplified a lot the code by using the ChangeHook.PostEvent that use the branch and do not require any change. Let me know what you think: https://gerrit-review.googlesource.com/#/c/50862

Patch-set: 4