Delete the ReplicationType enum, simplify the ReplicationState class
The ReplicationType parameter in the constructor of the ReplicationState class
was only used as a switch to find out which subclass of the
PushResultProcessing to instantiate. After that it was not used.  It could be
difficult for the ReplicationState class to provide all parameters needed to
instantiate any subclass of the PushResultProcessing. This becomes obvious from
[1] (lines 52 and 58).

Require users of the ReplicationState to pass an instance of the
PushResultProcessing.  This is more natural as the caller would already have
all parameters needed to instantiate that specific subtype of the
PushResultProcessing.

[1] https://gerrit-review.googlesource.com/#/c/50920/1/src/main/java/com/googlesource/gerrit/plugins/replication/ReplicationState.java

Change-Id: I945090bf1aa2bd9a2b3b2f00c5abc49f11624483