Update patch set 1
Patch Set 1:

> Is this fix worth putting on stable-2.8?

If you would like to have the change 50920 in stable-2.8 then it is worth because it would simplify the code in that change.
Otherwise, just for itself, it is not worth putting in stable-2.8.

Patch-set: 1
Reviewer: Saša Živkov <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0