Add method onRefReplicatedToAllNodes to PushResultProcessing
PushResultProcessing had 2 methods:

  onOneNodeReplicated->called when one ref is done replicating to a specific node
  onAllNodesReplicated->called when all refs are done replicating to all the nodes

It was missing a method for when replication is completed to all the nodes for one
ref. Add this method and renamed the 2 other methods for naming consistency:

  onRefReplicatedToOneNode->called when one ref is done replicating to a specific node
  onRefReplicatedToAllNodes->called when one ref is done replicating to all the nodes
  onAllRefsReplicatedToAllNodes->called when all the refs are done replicating to all the nodes

Sending stream event for ref replicated requires this change to prevent
implementing similar logic in all the implementations of PushResultProcessing.
Will simplify code in change 50920 and allow sending ref replicated stream
events for replication started by the command.

Change-Id: I89d162ea286c3ad20166843f5c3253662d9ae893