Update patch set 2
Patch Set 2: Code-Review+1

This will surely generate huge logs, but could certainly be useful in certain circumstances.

Regarding detecting replication issues:

Every now and then I run a script to help me verify that all refs on all slaves are up-to-date. I do this by generating sha1sums of the ls-remote output for all projects (on all slaves). (Using show-ref on the slaves FS would of course work fine too).

If the sha1sum of a ls-remote output for a project on a certain slave does not match the sha1sum of the masters ls-remote output, a more detailed report is generated for that specific project (listing all the refs that differs between the master and the slave). That way I can get decent speeds while still diffing millions of refs spread out on thousands of projects. And if the ls-remote differs in this first stage (perhaps due to an ongoing replication), it gets a second chance to catch up when the more detailed report is generated.

By putting the ls-remote processes in the background, I can fake some "multi-threading" in this ugly batch-script and gain some more speed.

Patch-set: 2
Reviewer: Gustaf Lundh <1006454@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1