Add --now switch to start replication right away
When starting replication from the command line, each replication task
is scheduled with the per remote replicationDelay.  This delay is
considered a batching mechanism for user driven events.  However,
scripts may desire to trigger replication without caring about batching.
In particular, a script may want to replicate a bunch of updates to many
projects serially for throttling reasons using the --wait option.
However, with a long list of projects to replicate the per project
replication delay can quickly add up to a lot of wasted time (too much
throttling).  By adding a --now option which starts replicating without
delay, it is possible for a script to control (generally along with the
--wait option) replication throttling very accurately.

Change-Id: I35b4386f53112ffa7c886ad345e5fabb97273aa9