Allow default force update to be configurable
In I4509d3500c730b8b5921dc155ea0476a70b3de9e the default refspec was
changed to not used forced update.

Administrators of sites that don't specify a refspec, because they
actually want replication to default to 'force push all refs', will
now need to edit their replication config to explicitly add the force
push refspec.

To make the configuration update a bit simpler, add a new setting
`gerrit.defaultForceUpdate` which, when set to true, will cause the
replication to behave in the same way as before when no refspec is
given.

Change-Id: Ia60a916135e87f9b7f36e5038c7595ed708045aa