Differentiate error logs for local and remote repository errors
RepositoryNotFoundException is raised when the local repository cannot
be opened.  RemoteRepositoryException is raised when there is an error
from the remote.

Both of these errors result in the same error message "Cannot replicate
project-name ..." being output to the log, which makes it difficult to
determine which of the errors has occurred.

Expand the log messages to make it possible to distinguish between local
and remote repository errors.

Change-Id: I63e24939200835e6c3a0bcacd4dd4c25ef766bf5