Fix: failure to create to missing remote repository via git://
When replicating to a remote via git:// and the remote repository
does not exist, it fails with RemoteRepositoryError rather than the
expected NoRemoteRepositoryException.  This means that the remote
repository never gets created, and the replication keeps getting
rescheduled.

Catch RemoteRepositoryErrors that are likely caused by the remote
repository not existing, and attempt to create it.

Bug: Issue 2420
Change-Id: Icf516481d5c2eccb520ba5f6673e3bac00f4db6f