Verify that created remote repository was actually created
When replicating to a remote on Gitlab, creating new projects fails
because Gitlab does not support it.  However due to a bug in the
Gitlab shell [1] it does not fail with an error code.  It is then
assumed that the creation was successful, and the replication is
rescheduled.  The same happens again next time, and the replication
keeps getting rescheduled.

After the repository creation, try to open the repository to verify
that it was actually created.  If the open fails, log an error and
do not reschedule the replication.

[1] https://github.com/gitlabhq/gitlab-shell/issues/127

Bug:  Issue 2421
Change-Id: I55a5a1852f1ccf558e0651fa7c8f77781747bb2a