Move replication logs into a separate file
Move the replication logs out of the Gerrit logs into a file called
replication_log.

Log statement for every replication is too verbose to be in the same
log file as the Gerrit logs. I also plan to print a PushOne identifier
in the logs so having it in a separate file will make this easier.

Change-Id: I3d2d3e3bbe8c155d171a81db9d72568ba5c1d0b0