Debug logging when replication is scheduled
It happens sometimes that replication is not correctly scheduled
when it is triggered by Gerrit events, but works fine when forced
through a replication SSH command.

Adding extra debug to actually see the replication schedule events
coming and then understand when they are not followed by a stateLog
replication info.

Change-Id: I9e04fd798c2a1ae07ba1a6515bff57926c4b3ca8