Prevent creating repos on extra servers
If using a group to replicate only certain repositories, it is possible
to be in a state where the authGroup is used on some servers but not
others.  If this happened, Gerrit would create the repository on all
servers, even if the authGroup would prevent replicating code to it.
By ensuring the authGroup can see the project first, we don't create
the repository if it's not needed.

Bug: Issue 543
Bug: Issue 2899
Change-Id: I2f189bdb6007228d0b25a6ce9c0357b3f97e4ced