Added regex and wildcard matching to ssh command
Patterns can now be used in the list of projects to replicate.
Actually queued projects are the intersection of patterns in
replication.config and the project list argument

This can be used e.g. when having large amounts of projects
where only those in a certain sub directory are to be replicated,
to avoid filling up the replication queue unnecessarily.

Change-Id: I44fa3fbfca5ee9067dd288e3ed45a256e1955d50