Update patch set 1
Patch Set 1:

Admins asked me for that feature, I am glad that you guys implemented it! I know it is not a bug fix but can we put that on stable-2.10 to get it as soon as possible, I would like to be able to use 2.10 replication plugin without forking it.

Patch-set: 1
Label: Verified=0