Make createMissingRepositories = false take affect on project-created event
Previously "createMissingRepositories = false" would prevent the replication
plugin from trying to create a new project when a ref-updated event was fired,
but when a project-created event was fired the replication plugin would try to
create a project on the remote.

"createMissingRepositories = false" should prevent all attempts of remote
project creation.

Change-Id: Ie616ee87de5d8e05cb99299a634ffcc6d50a0f32