Update patch set 2
Patch Set 2:

We want to use this setting to prevent repos from being automatically created on slaves that we use for partners.
We don't wish these partners to see all code. And even the names of some repos are restricted and should not be displayed to those who shouldn't see it.
Making this interpretation should also be valuable for other instances.
The fact is that in most cases you either want repos to be automatically created on remotes or you don't.
I can't really think of an instance where you want projects to be created on the remote if you have access to it when the project is created on the master, otherwise not.

Patch-set: 2