Update patch set 2
Patch Set 2:

(4 comments)

Apart from a few nits in the commit message it looks ok.

I can't really see the use-case though. Why would you want to use the replication plugin as a lib. What kind of functionality are you not able to implement in the replication plugin itself.

If there is a strong use-case maybe we should refactor the 'generic' replication parts out of the replication plugin and maintain them as a separate lib-project.

I'm pretty much clueless as to whether this is the right way forward or not. But from what I can see it should work :-)

Patch-set: 2
Label: Verified=0