Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

I'm not opposed to this -- but the easier solution is just not using a passphrase on the SSH key.  This doesn't add much security as Gerrit has to both be able to read the key and the passphrase stored in the config (though with the secure credential store, it may in the future)!

Patch-set: 3
Label: Code-Review=+1