Fix replication_log with external log4j.configuration
When using an external log4j configuration using the JVM container
option -Dlog4j.configuration=file:///... the replication log appender
wasn’t configured at all.

This was mainly due to the fact that all appenders were closed and 
removed before invoking the SystemLog.createAsyncAppender() that was
then unable to find any appender to attach.

Flipping the operations and using the single removeAppender instead
of the removeAllAppenders (the name is a lie: it does not only remove
but also *CLOSE* all the appenders, making them unusable) everything
is OK with an external configuration.

Issue: 3230
Change-Id: I4d9b153d49731a0a7d5c5df863c69dd2165cc883