Set connection timeout to 120 sec for SSH remote operations
The creation of a missing Git, before starting replication, is a
blocking operation. By setting a timeout on the synchronized
getSession(), we ensure the operation do not get stuck forever,
essentially blocking all future remote git creation operations.

120 seconds is an arbitrary limit, chosen because it should be enough
to not risk being triggered unnecessarily. This is a fail safe and we
should hopefully very rarely end up needing the timeout.

Change-Id: I01ed29004c6c75a75ede918f61fa1c68d246150b