Add new configuration option to control replication of hidden projects
In some cases we would like to replicate hidden projects to given
remotes. Right now it is not possible to achieve this, all hidden
projects all not replicated.

New remote.NAME.replicateHiddenProjects configuration allow to control
this behavior. By default it is set to 'false' to be consistent with
previous versions.

Change-Id: Ie7d7330ae685e104937037a89141cef2febbf5f1
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>