Add configuration option to replicate hidden projects
When projects are replicated to another Gerrit instance setting given
project as hidden will not have effect on the other side. This project
will be still visible on the 'slave' node. This cause data inconsistency
between 'master' and 'slave'

This is because replication plugin by default ignores hidden projects.
This patch adds configuration option that can change this behavior.

When remote.NAME.replicateHiddenPorojects is set to 'true' visibility
check will be ignored.

Change-Id: If8aab12dbf5d9b4dc7a9d051a43ab07890d97583
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>