Add config option to replicate not visible projects
When projects are replicated to another Gerrit instance setting given
project as hidden will not have effect on the other side. This project
will be still visible on the 'slave' node. This causes data inconsistency
between 'master' and 'slave'

This is because replication plugin by default ignores not visible projects.
This patch adds configuration option that can change this behavior.

When remote.NAME.replicateNotVisibleProjects is set to 'true' visibility
check will be ignored.

Change-Id: If8aab12dbf5d9b4dc7a9d051a43ab07890d97583
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>