ReplicationLogFile: remove unnecessary static 'started' flag
FindBugs warned about the static 'started' member being written to from
an instance method.

Further investigation revealed that the value of the variable is always
'false'.  This is because each plugin (re)load is done from a new class
loader, which means that a new ReplicationLogFile object is instantiated
and thus all its static members are initialized to their default value.

Remove it, since it's not actually doing what it was intended to do and
is thus unnecessary.

Change-Id: I50c0d929a2c422d46993e7ba7275edcc2d8def4f