Update patch set 1
Patch Set 1:

If core Gerrit's event dispatching methods are changed to have stricter parameters as in [1], the current implementation's event dispatching will not compile.

I'm not sure if this fix is the correct approach though.  Is there actually a reason to dispatch the event as a change event, even though it's not actually derived from ChangeEvent?

[1] https://gerrit-review.googlesource.com/62519

Patch-set: 1