Add method to push changes directly to given replica
This makes it possible to push changes directly to given replica
instance without sending unnecessary requests to others.

The method is intended to be used by other plugins that extend the
replication plugin.

This is a reworked version of the change that was originally done
on stable-2.16 but removed during the merge up to stable-3.1 in
change Iab801b94f.

Change-Id: Ib3e5a6c3afde834855d055b31beb43a560f8d785
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>
Signed-off-by: David Pursehouse <dpursehouse@digital.ai>