Update patch set 1
Patch Set 1:

I missed this earlier.
The AutoReloadConfigDecorator, responsible for loading config changes on the fly, only asks for the file from ReplicationBasedConfig, and if this file (etc/replication.config hasn't changed it doesn't reload the config.
This means that changes in the included files won't cause a config reload.
I guess a good way forward would be to make it ReplicationFileBasedConfig's responsibility to tell AutoReloadConfigDecorator when it was last changed, looping through the included files to check for changes.

Patch-set: 1
Label: Code-Review=0