Allow to configure timeout for SSH connection and commands
The timeout for SSH command execution is hard-coded to 0 which means
no timeout, and the timeout for SSH connections is hard-coded to 2
minutes.

Introduce new configuration options to allow configuring each timeout
individually:

  gerrit.sshCommandTimeout
  gerrit.sshConnectionTimeout

When not set, both default to their previously hard-coded values,
i.e. no limit and 2 minutes respectively.

Change-Id: Ibd377d45543ef4631082e8d521aeeb99209003f7
Signed-off-by: Dariusz Luksza <dluksza@collab.net>
Signed-off-by: David Pursehouse <dpursehouse@collab.net>