Handle kill command in replication tasks also while running
Issue description:
When replication tasks is scheduled and it gets killed with ssh kill
command further replication attempts get ignored and as a result
replication to particular mirror is no longer possible (Gerrit needs
to be restarted or plugin reloaded).

Solution:
Step1: Derive PushOne from CanceledWhileRunning interface so that
it gets informed that it was canceled while operation was
in progress (which results in any resource operation failure).

Step2: check in resource failure catch if it occurred while task
was canceled. If that is the case don't reschedule.

Change-Id: I4882fe1334efedf060a5ff5fef89d9d33c88d75c
Signed-off-by: Jacek Centkowski <geminica.programs@gmail.com>