Create new projects asynchronously via the replication queue
In the project created handler, new projects are created on remotes
synchronously, which can cause performance problems when the creation
operation takes a long time on one or more remotes.

Instead, insert the head name into the replication queue for the new
project, so that during the replication operation (which is executed
asynchronously) the project will be created.

Note that this causes a slight change in behavior: projects will never
be created on the remote if the createMissingRepositories setting is
disabled. Before this change, the project would always be created on
remotes during the intial creation (for remotes that are online), and
then any projects that were missing during actual replication would be
created if createMissingRepositories was enabled.

The default setting of createMissingRepositories is enabled, so it is
expected that the impact of this behavior change will be minimal.

Bug: Issue 3625
Change-Id: If429acda066e6b681918aa9268eda484fa5adf94