Create new projects asynchronously via the replication queue
In the project created handler, new projects are created on remotes
synchronously, which can cause performance problems when the creation
operation takes a long time on one or more remotes.

Instead, insert the head name into the replication queue for the new
project, so that during the replication operation (which is executed
asynchronously) the project will be created.

Bug: Issue 3625
Change-Id: If429acda066e6b681918aa9268eda484fa5adf94