Destination: parse replication delay and retry as time units
The replication delay and retry values are interpreted as seconds and
minutes respectively, but are being parsed as integers.

This is inconsistent with how time units are handled in other Gerrit
confiuration settings, and can cause confusion when the user configures
them using the time unit syntax such as "15s" and it causes the plugin
to fail with "invalid value".

Change it so that the delay and retry are parsed as time units. The
value can be given in any recognised time unit. Defaults remain the
same as before - 15 seconds and 1 minute respectively.

Change-Id: I7e518c6b8cac7f7af35d66df33eb6fd85d924eea