Update patch set 1
Patch Set 1:

> If it's was creating the repositories even when the option was
 > false, isn't that a bug?

No, createMissingRepositories options meant: while replicating, if the repo is not there(because replication missed the onNewProjectCreated event), create it.

I am ok with this change but we should update the config.md documentation to make it clear that is it's set to false, new project will never be replicated.

Patch-set: 1
Label: Code-Review=0