Handle the case when push was canceled while running
Issue description: When push gets canceled while actual operation
is performed it gets rescheduled.

Solution:
Step1 - Derive PushOne from CanceledWhileRunning interface so that
it gets informed that it was canceled while operation was in
progress (which results in resource operation failure).

Step2 - check TransportException catch if it occurred after the
task in question was canceled. If that is the case don't reschedule.

Change-Id: I4882fe1334efedf060a5ff5fef89d9d33c88d75c
Depends-On: I2e44407f68ebcf0aa92e1e9e30e10f26feed7182
Signed-off-by: Jacek Centkowski <geminica.programs@gmail.com>