Replicate HEAD reference when replicating a project
Replication plugin replicates HEAD changes to target instances giving
such changes are done using the web UI. If for any reason the target
instance was down when doing such change, or if the change was manually
done (either by editing the HEAD file directly or using a git command),
this HEAD change was not propagated and the target instance kept the
outdated head. Pushing manual replication did not replicate HEAD to
targets, and thus, the HEAD reference had to be manually updated on the
target instance.

Replicate HEAD reference when replicating a project so that HEAD changes
are always replicated to target instances, either when replication is
triggered automatically or when it is done using the 'replication start'
ssh command.

Change-Id: Ibd9be085dd56bf80a27e2798513fd06592209704