Replication plugin does not replicate HEAD in manual replication
Replication plugin replicates HEAD changes to target instances giving
such changes are done using the web UI. If for any reason the target
instance was down when doing a HEAD change, this HEAD change was not
propagated successfully and  the target instance kept the outdated head.

Pushing manual replication did not replicate HEAD to targets, and thus,
the HEAD reference had to be manually updated on the target instance.

Always replicate HEAD so that HEAD changes are replicated in any case.

Change-Id: Ibd9be085dd56bf80a27e2798513fd06592209704