Update patch set 1
Patch Set 1:

> i'm not really a fan of just reverting this.
 > the root of the slowness should be properly identified and
 > documented.
 > if it is an inherent problem of the feature, then maybe it should
 > be made optional?

Let's revert this change until we find a proper fix that doesn't introduce performance degradation.

We (Ericsson) introduced the change this week-end on our productions servers and notice the same performance degradation reported by Remy, we had to revert this change also.

We will try to look the head replication issue as soon as possible.

Patch-set: 1