Update patch set 1
Patch Set 1:

> Let's revert this change until we find a proper fix that doesn't
 > introduce performance degradation.
 > 
 > We will try to look the head replication issue as soon as possible.

I was planning to make 2.11.6 within the next week or so. We can leave this change under review for a while. If you can fix it on top of the current head of stable-2.11 then this can be abandoned; otherwise we can revert and then the better fix can go to stable-2.12 or master.

Patch-set: 1