Buck: Clean up java_test rule
We are maintaining two plugin artifacts: API and TEST_API where all
needed dependencies for compilation and running tests are exposed.
To make both plugin build modes work: in tree and standalone, we are
maintaining two variables, that are differently defined in different
build modes:

in tree:

  GERRIT_PLUGIN_API = ['//gerrit-plugin-api:lib']
  GERRIT_TESTS = ['//gerrit-acceptance-framework:lib']

standalone:

  GERRIT_PLUGIN_API = ['//lib/gerrit:plugin-api']
  GERRIT_TESTS = ['//lib/gerrit:acceptance-framework']

Given that replication plugin is a core plugin, we don't bother to use
the variables here, but use the in tree dependencies directly instead.

Change-Id: If7af37426ae0a7bfd49c48fd6276d666a31fe9ac