WiP: Add a new ssh command to abort replications
The command can be used to abort replications to one or more
remotes.

In this initial version, all replications to matching remotes are
aborted. In future it could be possible to abort only replication
of given projects, or by other criteria.

In dry run mode the command only lists the remotes that would be
aborted.

Change-Id: I173082414fca0556d5c0bcfe8a26ed7486de94d6