Update patch set 1
Patch Set 1:

Verified +1.
We at Sony Mobile have tested this patch with the environment of one Gerrit master(6 CPU cores and 94GB RAM) and four Gerrit slaves all running 2.12.3. One of the slaves is having limited access to some protected repositories and none of the changes submitted in these protected repositories are replicated this slave.

Patch-set: 1